requirements.txt
pandas 
numpy
seaborn
yfinance
python-dateutil
openpyxl
plotly
-e .